+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Add `include_existing_files` parameter to `ReplicationSetupHelper`
 
+### Fixed
+* Fix `AccountInfo.is_master_key()`
+
+### Infrastructure
+* Add 3.11.0-beta.1 to CI
+
 ## [1.16.0] - 2022-04-27
 
 This release contains a preview of replication support. It allows for basic